Fixed Akka non-terminating bugs
DeepDive have suffered from non-terminating bugs. These types of bugs
are because exceptions are thrown out of DeepDive.run, but the Akka
actor is not shut down. We have been trying to capture these bugs by
catching them everywhere and run a Context.shutdown, but the universal
solution is catching any type of exception and shut down the Akka actor
whenever an exception occurs.

Added a try-catch block in DeepDive.run, and shut down the system when
exceptions are thrown. This fixes config parsing failures and
unpredictable exceptions.