Drop a temporary column in grounding if it exists
During grounding, the 'dd_variable_type' column is added to the tables
involved in inference rules. If for some reason grounding gets killed,
the table is still there, so when running the application again, it
fails during grounding because the column already exists, so the
following query will fail:
'ALTER TABLE some_variable_table ADD COLUMN dd_variable_type int'

Now we try to remove the column before adding it.