Brings back multinomial with dense weight representation
which avoids incurring the extra (cross product, distinct, then join)
computation that can be very costly with categorical variable and
multinomial factors that almost always consider all possible categories
or their combinations.  Currently relies on
`DEEPDIVE_GROUNDING_DENSE_MULTINOMIAL=true` to switch to this, but we
can automatically switch representations based on the density of
categorical variables, which will be lightweight to compute.