Use TSV, python2 again for dumping sharded data
In the sharded pipeline, we move data from PostgreSQL (which has the
JSON type) to Greenplum. That process cleans up the data and makes it
uniformly JSON-compatible. This diff updates the pgtsv_to_json module
for dumping from Greenplum to HDFS.

It includes fixes for:

1. Reading escaped unicode from Greenplum as loaded from Postgres
(\u....),
2. Writing UTF-8-encoded JSON with \n-encoded newlines,
3. Handling "double precision" columns, and
4. Handling NULL non-text valued columns, and NULL values in non-text
array columns (derived from #550)